Rename dimension to dimensions
