+++ src/main/scala/libra/dimensions/All.scala
@@ -1,0 +1,4 @@
+package libra
+package dimensions
+
+trait All extends Angle
